Tested decorators: adapted and extended tests to new implementation.
